+++ ui/mods/com.pa.quitch.gwaioverhaul/gw_start/setup.js
@@ -85,7 +85,8 @@ if (!gwoSetupLoaded) {
             { id: "tgw_start_tank" },
             { id: "gwaio_start_nomad" },
             { id: "gwaio_start_backpacker" },
-            { id: "gwaio_start_hoarder" }
+            { id: "gwaio_start_hoarder" },
+            { id: "gwaio_start_warp" }
           );
           if (!model.gwoStartingCards) {
             model.gwoStartingCards = [];